% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postStats.R
\name{EstimatePosteriorStats}
\alias{EstimatePosteriorStats}
\title{Estimate posterior stats and Monte Carlo errors}
\usage{
EstimatePosteriorStats(mc, burnin = 1)
}
\arguments{
\item{mc}{Markov chain}

\item{burnin}{number of initial elements of Markov chain to discard}
}
\description{
Estimate posterior stats and Monte Carlo errors
}
\keyword{internal}
