% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiteDNAoccModel.R
\name{fitOccModel}
\alias{fitOccModel}
\title{Fit a Multi-scale Occupancy Model}
\usage{
fitOccModel(niter = 100, niterInterval = 10, y, K, X, W, V, siteEffectInW,
  colNamesOfX, colNamesOfW, colNamesOfV)
}
\arguments{
\item{niter}{no. iterations of MCMC algorithm}

\item{niterInterval}{no. iterations for reporting progress of MCMC algorithm}

\item{y}{M x J matrix of numbers of detections per sample}

\item{K}{M x J matrix of numbers of replicates per sample}

\item{X}{matrix of regressors associated with model parameter beta}

\item{W}{array of regressors associated with model parameter alpha}

\item{V}{array of regressors associated with model parameter delta}

\item{siteEffectInW}{logical indicator of whether model parameter alpha contains M elements (that is, one element per site)}

\item{colNamesOfX}{vector of names of regressors in X}

\item{colNamesOfW}{vector of names of regressors in W}

\item{colNamesOfV}{vector of names of regressors in V}
}
\description{
This function is called by \code{occModel} -- rather than directly by the user -- and fits a multi-scale occupancy model using Bayesian methods (specifically, a Markov chain Monte Carlo (MCMC) algorithm).
}
\keyword{internal}
