% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungusSurveyData.R
\docType{data}
\name{fungusSurveyData}
\alias{fungusSurveyData}
\title{Fungal Pathogen Survey Data}
\format{A data frame consisting of 20 rows and the following 5 columns:
\describe{
\item{site:}{ location sampled for fungal pathogen }
\item{loadT1:}{ average fungal pathogen load of frogs that tested positive for Bd and were captured during 1st survey }
\item{loadT2:}{ average fungal pathogen load of frogs that tested positive for Bd and were captured during 2nd survey }
\item{frogs:}{ index of frog density (catch per unit effort) }
\item{Bd:}{ prevalence of Bd among frogs captured during 1st or 2nd surveys }
}}
\source{
Schmidt BR, Kery M, Ursenbacher S, Hyman OJ, and Collins JP (2013) Site occupancy models in the analysis of environmental DNA presence/absence surveys: a case study of an emerging amphibian pathogen.  Methods in Ecology and Evolution 4: 646-653.
}
\usage{
data(fungusSurveyData)
}
\description{
A data set containing environmental and ecological variables measured at each of 20 ponds located in Coconino National Forest, Arizona, USA.  At each site, water samples were collected to extract the DNA of \emph{Batrachochytrium dendrobatidis} (Bd), a fungal pathogen of various species of amphibians.
}
\references{
Schmidt BR, Kery M, Ursenbacher S, Hyman OJ, and Collins JP (2013) Site occupancy models in the analysis of environmental DNA presence/absence surveys: a case study of an emerging amphibian pathogen.  Methods in Ecology and Evolution 4: 646-653.

\url{http://dx.doi.org/10.1111/2041-210X.12052}

Hyman, OJ, Collins JP (2012) Evaluation of a filtration-based method for detecting  \emph{Batrachochytrium dendrobatidis} in natural bodies of water.  Diseases of Aquatic Organisms 97: 185-195.

\url{ https://doi.org/10.3354/dao02423}
}
\keyword{datasets}
