% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occData.R
\name{occData}
\alias{occData}
\title{Compute Occupancy Data Matrices}
\usage{
occData(d, siteColName = "site", sampleColName = "sample")
}
\arguments{
\item{d}{data frame containing tabular detection data}

\item{siteColName}{column of data frame d containing sites (i.e., primary sample units)}

\item{sampleColName}{column of data frame d containing sample numbers (i.e., integers 1, 2, ...) for each site}
}
\value{
A list containing two matrices:
\describe{
\item{y:}{   M x J matrix of numbers of detections per sample}
\item{K:}{   M x J matrix of numbers of replicates per sample}
}
where M = no. sites and J = maximum no. samples per site
}
\description{
Computes occupancy data matrices from a data frame of tabular detection data.
}
\details{
This function takes a data frame of tabular detection data and returns a list containing a matrix of numbers of detections per sample and a matrix of numbers of replicates per sample.

The number of samples per site may differ among sites.  In addition, the number of replicates per sample may differ among samples.  Such unbalanced designs will induce NAs in the detection matrix (y) and zeros in the replicate number matrix (K).
}
\examples{
data(gobyDetectionData)

occData(gobyDetectionData, "site", "sample")
}
