% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occModel.R
\name{occModel}
\alias{occModel}
\title{Specify and Fit a Multi-scale Occupancy Model}
\usage{
occModel(formulaSite = ~1, formulaSiteAndSample = ~1,
  formulaReplicate = ~1, detectionMats, siteData = NULL,
  siteAndSampleData = NULL, niter = 1100, niterInterval = 100,
  siteColName = "site", sampleColName = "sample")
}
\arguments{
\item{formulaSite}{model of occurrence at a site}

\item{formulaSiteAndSample}{model of occurrence in different samples of a site}

\item{formulaReplicate}{model of detection in replicates of a sample}

\item{detectionMats}{list containing
\describe{
\item{y:}{   M x J matrix of numbers of detections per sample}
\item{K:}{   M x J matrix of numbers of replicates per sample}
}
where M = no. sites and J = maximum no. samples per site}

\item{siteData}{data frame containing site-level covariates}

\item{siteAndSampleData}{data frame containing site- and sample-level covariates}

\item{niter}{no. iterations of MCMC algorithm}

\item{niterInterval}{no. iterations for reporting progress of MCMC algorithm}

\item{siteColName}{column of data frame siteData (or column of data frame siteAndSampleData) containing sites (i.e., primary sample units)}

\item{sampleColName}{column of data frame siteData (or column of data frame siteAndSampleData) containing sample numbers (i.e., integers 1, 2, ...) for each site}
}
\value{
A list (object of class occModel) containing the following objects:
\describe{
\item{niterations:}{total number of iterations MCMC algorithm was run}
\item{state:}{the value of each model parameter after the last iteration of the MCMC algorithm}
\item{y:}{  M x J matrix of numbers of detections per sample}
\item{K:}{  M x J matrix of numbers of replicates per sample}
\item{X:}{ matrix of regressors associated with model parameter beta}
\item{W:}{ array of regressors associated with model parameter alpha}
\item{V:}{ array of regressors associated with model parameter delta}
\item{siteEffectInW:}{ logical indicator of whether model parameter alpha contains M elements (that is, one element per site)}
\item{colNamesOfX:}{ vector of names of regressors in X}
\item{colNamesOfW:}{ vector of names of regressors in W}
\item{colNamesOfV:}{ vector of names of regressors in V}
}

where M = no. sites and J = maximum no. samples per site
}
\description{
This function specifies and fits a multi-scale occupancy model.
}
\details{
This function is used to specify and fit a multi-scale occupancy model (with or without covariates).  The model is fitted using Bayesian methods (specifically, a Markov chain Monte Carlo (MCMC) algorithm that is run for a finite number of iterations).  Additional MCMC iterations may be computed using the function \code{updateOccModel}.   Output from the MCMC algorithm is stored in the file "mc.csv" and can be summarized using functions \code{plotTrace}, \code{plotACF}, and \code{posteriorSummary}.
}
\note{
Before fitting the model, this function checks to ensure that the model specification is possible given the data files.  These checks include:
\itemize{
\item  Value of siteColName matches that in data frame siteData.
\item  Values of siteColName and sampleColName match those in data frame siteAndSampleData.
\item  Values in sampleColName contain consecutive positive integers (1, 2, ...) for each site.
\item  Only one data frame (siteData or siteAndSampleData) is specified for survey data.
}

If any of these checks fail, the function returns an error message.
}
\examples{
data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")

# Fit a model without covariates of occurrence or detection.
fit.simplest = occModel(detectionMats=detections)


# Fit a model which assumes that probability of occurrence in samples differs with site,
# but the differences are not a function of site-level covariates.

data(gobySurveyData)
fit = occModel(~1, ~factor(site)-1, ~1,
          detectionMats=detections,
          siteData=gobySurveyData,
          siteColName="site")


# Fit a model assuming occurrence and detection probabilities
# are functions of site-level covariates.

data(gobySurveyData)
gobySurveyData = scaleData(gobySurveyData)  # center and scale numeric covariates
fit1 = occModel(formulaSite          = ~ veg,
                formulaSiteAndSample = ~ sal + twg,
                formulaReplicate     = ~ sal + fish,
                detectionMats        = detections,
                siteData             = gobySurveyData,
                niter                = 110,
                niterInterval        = 10,
                siteColName = 'site',
                )


# Update the Markov chain of the model specified in fit1
fit2 = updateOccModel(fit1, niter=50, niterInterval=10)
}
\seealso{
\code{\link{updateOccModel}}, \code{\link{scaleData}}
}
