% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACF.R
\name{plotACF}
\alias{plotACF}
\title{Plot ACF for One or More Parameters of a Multi-scale Occupancy Model}
\usage{
plotACF(fit, paramName, burnin = 1)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{paramName}{vector of names of parameters to be plotted}

\item{burnin}{initial no. iterations of Markov chain to be omitted from plot}
}
\description{
Plots the ACF (autocorrelation function) of the Markov chain for each of 1-4 parameters of a multi-scale occupancy model.
}
\details{
This function plots the ACF (autocorrelation function) of the Markov chain created by \code{occModel} for each of 1-4 parameters of a multi-scale occupancy model.
}
\examples{
data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
fit = occModel(detectionMats=detections)
plotACF(fit, c('beta..Intercept.', 'alpha..Intercept.', 'delta..Intercept.'), burnin=100)
}
