% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{plotTrace}
\alias{plotTrace}
\title{Plot Trace of One or More Parameters of a Multi-scale Occupancy Model}
\usage{
plotTrace(fit, paramName, burnin = 1)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{paramName}{vector of names of parameters to be plotted}

\item{burnin}{initial no. iterations of Markov chain to be omitted from plot}
}
\description{
Plots the Markov chain for each of 1-4 parameters of a multi-scale occupancy model.
}
\details{
This function plots the Markov chain created by \code{occModel} for each of 1-4 parameters of a multi-scale occupancy model.
}
\examples{
data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
fit = occModel(detectionMats=detections)
plotTrace(fit, 'beta.(Intercept)', burnin=100)
plotTrace(fit, c('beta.(Intercept)', 'alpha.(Intercept)', 'delta.(Intercept)'), burnin=100)
}
