% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSummary.R
\name{posteriorSummary}
\alias{posteriorSummary}
\title{Posterior Summary For Parameters of a Multi-scale Occupancy Model}
\usage{
posteriorSummary(fit, burnin = 1, mcError = FALSE,
  outputSummary = FALSE)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{burnin}{initial no. iterations of Markov chain to be omitted from calculations}

\item{mcError}{logical switch to estimate Monte Carlo standard errors}

\item{outputSummary}{logical switch to return values of posterior summary statistics}
}
\value{
Prints estimates of posterior summary statistics.  If \code{mcError}=TRUE, the Monte Carlo standard errors of these estimates are computed.   If \code{outputSummary}=TRUE, the posterior summaries are returned in a list.
}
\description{
Estimates the posterior mean, median, and 95\% credible limits for each parameter of a multi-scale occupancy model.
}
\details{
This function estimates the posterior mean, median, and 95\% credible limits for each parameter of a multi-scale occupany model using a Markov chain created by \code{occModel}.
}
\examples{
data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
fit = occModel(detectionMats=detections)
posteriorSummary(fit, burnin=100,  mcError = TRUE)
}
