% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSummaryOfPsi.R
\name{posteriorSummaryOfSiteOccupancy}
\alias{posteriorSummaryOfSiteOccupancy}
\title{Posterior Summary of a Multi-scale Occupancy Model's Site-Occupancy Probabilities}
\usage{
posteriorSummaryOfSiteOccupancy(fit, burnin = 1, mcError = FALSE)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{burnin}{initial no. iterations of Markov chain to be omitted from calculations}

\item{mcError}{logical switch to estimate Monte Carlo standard errors}
}
\value{
Computes estimates of summary statistics for a multi-scale occupancy model's site-occupancy probabilities.  If \code{mcError}=TRUE, the Monte Carlo standard errors of these estimates are computed.   All posterior summaries are returned in a list.
}
\description{
Estimates the posterior mean, median, and 95\% credible limits for a multi-scale occupancy model's site-specific occurrence probabilities.
}
\details{
This function estimates the posterior mean, median, and 95\% credible limits of the site-specific occurrence probabilities of a multi-scale occupancy model.
}
\examples{

data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
data(gobySurveyData)
gobySurveyData = scaleData(gobySurveyData)  # center and scale numeric covariates

fit1 = occModel(formulaSite          = ~ veg,
                formulaSiteAndSample = ~ sal + twg,
                formulaReplicate     = ~ sal + fish,
                detectionMats        = detections,
                siteData             = gobySurveyData,
                niter                = 1100,
                niterInterval        = 100,
                siteColName = 'site',
                )

psi = posteriorSummaryOfSiteOccupancy(fit1, burnin=100)
plot(gobySurveyData[, 'turb'], psi$median)
}
