% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleData.R
\name{scaleData}
\alias{scaleData}
\title{Center and Scale Numeric Columns of a Data Frame}
\usage{
scaleData(df, sampleColName = NULL)
}
\arguments{
\item{df}{a data frame object}

\item{sampleColName}{column of data frame containing sample numbers (i.e., integers 1, 2, ...) for each site}
}
\value{
The input data frame with all numeric- or integer-valued columns transformed to have mean zero and unit variance.
}
\description{
Centers and scales all numeric- or integer-valued columns of a data frame object.
}
\details{
This function extracts all numeric- or integer-valued columns of a data frame object.  Each of these columns is centered by subtracting its mean and then scaled by dividing the result by the standard deviation of the original column values.  This linear transformation produces columns with mean zero and unit variance.  Both input and output data frame objects have the same column ordering.  sampleColName is only needed if data frame contains sample numbers.
}
