% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateeDNAoccModel.R
\name{updateOccModel}
\alias{updateOccModel}
\title{Update a Multi-scale Occupancy Model}
\usage{
updateOccModel(fit, niter, niterInterval)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{niter}{no. iterations of MCMC algorithm}

\item{niterInterval}{no. iterations for reporting progress of MCMC algorithm}
}
\value{
A list (object of class occModel) containing the following objects:
\describe{
\item{niterations:}{total number of iterations MCMC algorithm was run}
\item{state:}{the value of each model parameter after the last iteration of the MCMC algorithm}
\item{y:}{  M x J matrix of numbers of detections per sample}
\item{K:}{  M x J matrix of numbers of replicates per sample}
\item{X:}{ matrix of regressors associated with model parameter beta}
\item{W:}{ array of regressors associated with model parameter alpha}
\item{V:}{ array of regressors associated with model parameter delta}
\item{siteEffectInW:}{ logical indicator of whether model parameter alpha contains M elements (that is, one element per site)}
\item{colNamesOfX:}{ vector of names of regressors in X}
\item{colNamesOfW:}{ vector of names of regressors in W}
\item{colNamesOfV:}{ vector of names of regressors in V}
}

where M = no. sites and J = maximum no. samples per site
}
\description{
This function computes additional iterations of a Markov chain Monte Carlo algorithm that was used to fit a multi-scale occupancy model to a data set.
}
\details{
This function is used to update a multi-scale occupancy model fitted using \code{occModel}.  The model is fitted using Bayesian methods (specifically, a Markov chain Monte Carlo (MCMC) algorithm that is run for a finite number of iterations).  Output from the MCMC algorithm is stored in the file "mc.csv" and can be summarized using functions \code{plotTrace}, \code{plotACF}, and \code{posteriorSummary}.
}
\examples{
data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
data(gobySurveyData)
gobySurveyData = scaleData(gobySurveyData)  # center and scale numeric covariates

fit1 = occModel(formulaSite          = ~ veg,
                formulaSiteAndSample = ~ sal + twg,
                formulaReplicate     = ~ sal + fish,
                detectionMats        = detections,
                siteData             = gobySurveyData,
                niter                = 110,
                niterInterval        = 10,
                siteColName = 'site',
                )

# Update the Markov chain of the model specified in fit1
fit2 = updateOccModel(fit1, niter=50, niterInterval=10)

}
\seealso{
\code{\link{occModel}}, \code{\link{scaleData}}
}
