% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAIC.R
\name{WAIC}
\alias{WAIC}
\title{Widely Applicable Information Criterion (WAIC) of a Multi-scale Occupancy Model}
\usage{
WAIC(fit, burnin = 1)
}
\arguments{
\item{fit}{object of class occModel that contains data and previous state of the model's Markov chain}

\item{burnin}{initial no. iterations of Markov chain to be omitted from calculations}
}
\value{
value of WAIC and its goodness-of-fit and predictive-variance components.
}
\description{
Computes the value of WAIC for a fitted multi-scale occupancy model.
}
\details{
This function computes the WAIC value used in model-selection once a multi-scale occupancy model has been fitted using  \code{occModel}.
}
\examples{

data(gobyDetectionData)
detections = occData(gobyDetectionData, "site", "sample")
data(gobySurveyData)
gobySurveyData = scaleData(gobySurveyData)  # center and scale numeric covariates

fit1 = occModel(formulaSite          = ~ veg,
                formulaSiteAndSample = ~ sal + twg,
                formulaReplicate     = ~ sal + fish,
                detectionMats        = detections,
                siteData             = gobySurveyData,
                niter                = 1100,
                niterInterval        = 100,
                siteColName = 'site',
                )

WAIC(fit1, burnin=100)
}
